Publish a new version, v15.3.0

+++ localeval.js
@@ -132,7 +132,9 @@ if (node_js) {
   };
 
   evaluator.clear = function() {
-    child.kill('SIGKILL');
+    if (child) {
+      child.kill('SIGKILL');
+    }
   };
 
   return evaluator;
